Max patch that extracts time from transcription
This Max patch listens to OSC messages sent from the Python script, and waits for a message with the word "reminder", and uses some convoluted methods to extract an integer from the time spoken and am/pm keyword. Of course this would be trivial in Python, but serves as an example for those using Max.